only owner can manage to its data

+++ app/Http/Controllers/ContactsController.php
@@ -7,33 +7,50 @@ use Illuminate\Http\Request;
 
 class ContactsController extends Controller
 {
-    public function store()
-    {
-      Contact::create($this->validateData());
+  public function index()
+  {
+    return request()->user()->contacts;
+  }
+
+  public function store()
+  {
+    request()->user()->contacts()->create($this->validateData());
+  }
+
+  public function show(Contact $contact)
+  {
+    if (request()->user()->isNot($contact->user)) {
+      return response([], 403);
     }
 
-    public function show(Contact $contact)
-    {
-      return $contact;
+    return $contact;
+  }
+
+  public function update(Contact $contact)
+  {
+    if (request()->user()->isNot($contact->user)) {
+      return response([], 403);
     }
 
-    public function update(Contact $contact)
-    {
-      $contact->update($this->validateData());
+    $contact->update($this->validateData());
+  }
+
+  public function destroy(Contact $contact)
+  {
+    if (request()->user()->isNot($contact->user)) {
+      return response([], 403);
     }
 
-    public function destroy(Contact $contact)
-    {
-      $contact->delete();
-    }
+    $contact->delete();
+  }
 
-    private function validateData()
-    {
-      return request()->validate([
-        'name' => 'required',
-        'email' => 'required|email',
-        'birthday' => 'required',
-        'company' => 'required',
-      ]);
-    }
+  private function validateData()
+  {
+    return request()->validate([
+      'name' => 'required',
+      'email' => 'required|email',
+      'birthday' => 'required',
+      'company' => 'required',
+    ]);
+  }
 }
